requirements.txt
web3
python-dotenv
discord-webhook
asyncio
pytest